fix: review fixes and refactors
