+++ src/components/popups/ChangeTierPopup.vue
@@ -2,13 +2,9 @@
 import { ref, PropType, onMounted } from 'vue';
 import Avatar from '@/components/Avatar.vue';
 import CloseIcon from '@/components/icons/XIcon.vue';
-import CrownIcon from '@/components/icons/Crown.vue';
-import CheckCircleIcon from '@/components/icons/CheckCircle.vue';
-import ChevronDownIcon from '@/components/icons/ChevronDown.vue';
-import SpinnerIcon from '@/components/icons/SpinnerIcon.vue';
-import SwitchPeriod from '@/components/ToggleSwitch.vue';
+import EditTierSelect from '@/components/EditTierSelect.vue';
+import EditTierConfirm from '../EditTierConfirm.vue';
 import { useStore } from '@/store/session';
-import { useSubscriptionStore } from '@/store/subscriptions';
 import { ISubscriptionWithProfile } from '@/store/subscriptions';
 import {
 	usePaymentsStore,
@@ -16,8 +12,8 @@ import {
 	PaymentProfile,
 	createDefaultPaymentProfile,
 } from '@/store/paymentProfile';
-import { canSwitchSubscription, getCurrencySymbol } from '@/backend/payment';
-import { toastError, toastSuccess, handleError } from '@/plugins/toast';
+import { canSwitchSubscription } from '@/backend/payment';
+import { toastError, handleError } from '@/plugins/toast';
 import { Profile } from '@/backend/profile';
 
 const props = defineProps({
@@ -46,17 +42,14 @@ const props = defineProps({
 });
 
 const store = useStore();
-const useSubscription = useSubscriptionStore();
 const usePayment = usePaymentsStore();
 const step = ref<number>(0);
-const selectedTier = ref<SubscriptionTier | null>(null);
+const selectedTier = ref<SubscriptionTier>(props.toPreSelectTier);
 const selectedPeriod = ref<string>(`month`);
 const paymentProfile = ref<PaymentProfile>(createDefaultPaymentProfile(store.$state.id));
-const isLoading = ref<boolean>(false);
-// TODO:will update when canSwitchSubscription is available. set to true for testing
 const canSwitchTier = ref<boolean>(true);
 
-const emit = defineEmits([`close`]);
+defineEmits([`close`]);
 
 onMounted(async (): Promise<void> => {
 	usePayment.getPaymentProfile(props.author.id);
@@ -69,14 +62,11 @@ onMounted(async (): Promise<void> => {
 		canSwitchTier.value = canSwitchResponse;
 		return;
 	} catch (err) {
-		throw err;
+		handleError(err);
 	}
 });
 
 // methods
-function displayCurrency(currency: string): string {
-	return getCurrencySymbol(currency);
-}
 async function initializeProfile(): Promise<void> {
 	if (!props.author) {
 		toastError(`Author profile is missing`);
@@ -99,59 +89,10 @@ async function initializeProfile(): Promise<void> {
 		toastError(`Author hasn't set-up subscriptions`);
 	}
 }
-function selectTier(tier: SubscriptionTier): void {
-	selectedTier.value = tier;
-}
-function switchPeriod(): void {
-	if (selectedPeriod.value === `month`) {
-		selectedPeriod.value = `year`;
-	} else {
-		selectedPeriod.value = `month`;
-	}
-}
-function nextStep(): void {
+function nextStep(selectTier: SubscriptionTier, selectPeriod: string): void {
+	selectedTier.value = selectTier;
+	selectedPeriod.value = selectPeriod;
 	step.value += 1;
-}
-async function switchTier(): Promise<void> {
-	if (selectedTier.value === null) {
-		return;
-	}
-	try {
-		isLoading.value = true;
-		const res = await useSubscription.switchSubscriptionTier(
-			store.$state.id,
-			props.sub.subscriptionId,
-			selectedTier.value,
-			selectedPeriod.value,
-		);
-		if (!res) {
-			toastError(`Switching tier failed`);
-			return;
-		} else {
-			toastSuccess(`Switched tiers successfully!`);
-			emit(`close`);
-		}
-	} catch (err: any) {
-		handleError(err);
-	}
-	isLoading.value = false;
-}
-function getStyles(DisplayedTier: SubscriptionTier): string {
-	let res = ``;
-	if (props.sub.tier.id === DisplayedTier._id) {
-		// current tier
-		res = `opacity-75 cursor-not-allowed border-gray5`;
-	} else if (props.enabledTiers.length > 0 && !props.enabledTiers.includes(DisplayedTier._id)) {
-		// not in enabled tiers for this post
-		res = `opacity-75 cursor-not-allowed border-lightBorder dark:border-darkBorder`;
-	} else if (selectedTier.value?._id === DisplayedTier._id) {
-		// in enabled tier and selected
-		res = `opacity-100 cursor-pointer border-neutral`;
-	} else {
-		// in enabled tier but not selected
-		res = `opacity-100 cursor-pointer border-lightBorder dark:border-darkBorder`;
-	}
-	return res;
 }
 
 initializeProfile();
@@ -193,186 +134,25 @@ initializeProfile();
 					</button>
 				</div>
 				<!-- Step 0: Change tier -->
-				<article v-show="step === 0">
-					<div class="w-full flex flex-col justify-center text-center px-10">
-						<CrownIcon class="text-neutral stroke-neutral self-center w-12 h-12 mb-2" />
-						<h6 class="font-semibold text-neutral text-xl mb-2">Change Tier</h6>
-						<p v-if="canSwitchTier" class="text-base text-center text-gray5 dark:text-gray3 mb-4">
-							Easily change your Tier access of
-							<span v-if="author.name !== ``" class="font-semibold text-primary dark:text-secondary">{{
-								author.name
-							}}</span>
-							<span v-else class="font-semibold text-primary dark:text-secondary">@{{ author.id }}</span>
-							to access new content
-						</p>
-					</div>
-					<div v-if="!canSwitchTier">
-						<p class="text-base text-center text-gray5 dark:text-gray3 mb-10 mt-2">
-							You already switched Tiers for this author within the last 30 days. <br />
-							You must wait for your next payment to process before changing the subscription tier again.
-						</p>
-						<div class="flex flex-row-reverse">
-							<button
-								class="bg-darkBG text-lightButtonText focus:outline-none transform rounded-lg font-bold transition duration-500 ease-in-out hover:bg-opacity-75"
-								style="padding: 0.4rem 1.5rem"
-								@click="$emit(`close`)"
-							>
-								<span class="font-sans" style="font-size: 0.95rem"> Close </span>
-							</button>
-						</div>
-					</div>
-					<div v-else>
-						<!-- Period switch -->
-						<div class="w-full flex justify-center mt-1">
-							<SwitchPeriod :period="selectedPeriod" @toggle="switchPeriod" />
-						</div>
-						<!-- Subscriptions list -->
-						<div v-for="tier in paymentProfile.tiers" :key="tier._id">
-							<button
-								v-if="
-									(tier.monthlyEnabled && selectedPeriod === `month`) ||
-									(tier.yearlyEnabled && selectedPeriod === `year`)
-								"
-								class="flex flex-row items-center justify-between m-5 p-4 border shadow-sm rounded-lg bg-lightBG dark:bg-darkBG transition duration-500 ease-in-out"
-								:class="getStyles(tier)"
-								:disabled="
-									props.sub.tier.id === tier._id || (enabledTiers.length > 0 && !enabledTiers.includes(tier._id))
-								"
-								@click="selectTier(tier)"
-							>
-								<!-- Check mark -->
-								<div class="w-12 flex justify-center">
-									<CheckCircleIcon
-										v-if="props.sub.tier.id !== tier._id"
-										:is-checked="selectedTier !== null && selectedTier._id === tier._id"
-										class="text-neutral w-6 h-6 flex items-center transition duration-500 ease-in-out"
-									/>
-									<CheckCircleIcon
-										v-else
-										:is-checked="true"
-										class="text-gray5 w-6 h-6 flex items-center transition duration-500 ease-in-out"
-									/>
-								</div>
-								<div class="flex flex-grow flex-col items-start ml-4 mr-2 w-2/5">
-									<h3 class="text-xl font-semibold dark:text-darkPrimaryText">{{ tier.name }}</h3>
-									<p class="text-gray5 dark:text-gray3 text-left text-sm pr-2">
-										Get access to exclusive articles by subscribing to {{ tier.name }}
-									</p>
-								</div>
-								<div v-if="sub.tier.id !== tier._id">
-									<div
-										v-if="tier.monthlyEnabled && selectedPeriod === `month`"
-										class="font-semibold text-lg mr-2 dark:text-darkPrimaryText"
-									>
-										{{ displayCurrency(paymentProfile.currency) }}{{ tier.monthlyPrice.toLocaleString() }}
-										<span class="text-gray5 dark:text-gray3">/month</span>
-									</div>
-									<div
-										v-if="tier.yearlyEnabled && selectedPeriod === `year`"
-										class="font-semibold text-lg mr-2 dark:text-darkPrimaryText"
-									>
-										{{ displayCurrency(paymentProfile.currency) }}{{ tier.yearlyPrice.toLocaleString() }}
-										<span class="text-gray5 dark:text-gray3">/year</span>
-									</div>
-								</div>
-								<div v-else class="font-semibold text-lg mr-2 text-gray5">Current tier</div>
-							</button>
-						</div>
-						<div class="flex flex-row-reverse">
-							<button
-								:class="selectedTier !== null ? `` : `opacity-50 cursor-not-allowed`"
-								class="bg-darkBG text-lightButtonText focus:outline-none transform rounded-lg font-bold transition duration-500 ease-in-out hover:bg-opacity-75"
-								style="padding: 0.4rem 1.5rem"
-								:disabled="selectedTier === null"
-								@click="nextStep"
-							>
-								<span class="font-sans" style="font-size: 0.95rem"> Next </span>
-							</button>
-						</div>
-					</div>
+				<article v-if="step === 0">
+					<EditTierSelect
+						:payment-profile="paymentProfile"
+						:can-switch-tier="canSwitchTier"
+						:sub="props.sub"
+						:author="props.author"
+						@next-step="nextStep"
+					/>
 				</article>
 				<!-- Step 1: Confirmation page -->
-				<article v-show="step === 1" class="flex flex-col items-center">
-					<div class="w-full flex flex-col justify-center text-center px-10">
-						<CrownIcon class="text-neutral stroke-neutral self-center w-12 h-12 mb-2" />
-						<h6 class="font-semibold text-neutral text-xl mb-2">Are you sure?</h6>
-						<p class="text-base text-center text-gray5 dark:text-gray3">
-							You are about to change the Tier subscription of author
-							<span v-if="author.name !== ``" class="font-semibold text-primary dark:text-secondary">{{
-								author.name
-							}}</span>
-							<span v-else class="font-semibold text-primary dark:text-secondary">@{{ author.id }}</span>
-						</p>
-					</div>
-					<!-- Tier change preview -->
-					<div
-						class="flex flex-row items-center justify-between m-5 p-4 border shadow-sm rounded-lg bg-lightBG dark:bg-darkBG transition duration-500 ease-in-out opacity-100 border-gray5"
-					>
-						<!-- Check mark -->
-						<div class="w-12 flex justify-center">
-							<CheckCircleIcon
-								:is-checked="true"
-								class="text-gray5 w-6 h-6 flex items-center transition duration-500 ease-in-out"
-							/>
-						</div>
-						<div class="flex flex-grow flex-col items-start ml-4 mr-2 w-2/5">
-							<h3 class="text-xl font-semibold dark:text-darkPrimaryText">{{ props.sub.tier.name }}</h3>
-							<p class="text-gray5 dark:text-gray3 text-left text-sm pr-2">
-								Get access to exclusive articles by subscribing to {{ props.sub.tier.name }}
-							</p>
-						</div>
-						<div class="font-semibold text-lg mr-2 text-gray5">Current tier</div>
-					</div>
-					<ChevronDownIcon class="w-5 h-5 text-gray5 dark:text-gray3" />
-					<div
-						class="flex flex-row items-center justify-between m-5 p-4 border shadow-sm rounded-lg bg-lightBG dark:bg-darkBG transition duration-500 ease-in-out opacity-100 border-neutral"
-					>
-						<!-- Check mark -->
-						<div class="w-12 flex justify-center">
-							<CheckCircleIcon
-								:is-checked="true"
-								class="text-neutral w-6 h-6 flex items-center transition duration-500 ease-in-out"
-							/>
-						</div>
-						<div v-if="selectedTier" class="flex flex-grow flex-col items-start ml-4 mr-2 w-2/5">
-							<h3 class="text-xl font-semibold dark:text-darkPrimaryText">{{ selectedTier.name }}</h3>
-							<p class="text-gray5 dark:text-gray3 text-left text-sm pr-2">
-								Get access to exclusive articles by subscribing to {{ selectedTier.name }}
-							</p>
-						</div>
-						<div
-							v-if="selectedTier && selectedPeriod === `month`"
-							class="font-semibold text-lg mr-2 dark:text-darkPrimaryText"
-						>
-							{{ displayCurrency(paymentProfile.currency) }}{{ selectedTier.monthlyPrice.toLocaleString() }}
-							<span class="text-gray5 dark:text-gray3">/month</span>
-						</div>
-						<div
-							v-if="selectedTier && selectedPeriod === `year`"
-							class="font-semibold text-lg mr-2 dark:text-darkPrimaryText"
-						>
-							{{ displayCurrency(paymentProfile.currency) }}{{ selectedTier.yearlyPrice.toLocaleString() }}
-							<span class="text-gray5 dark:text-gray3">/year</span>
-						</div>
-					</div>
-					<p class="text-sm text-center text-negative px-8 mb-5 py-1">
-						This change will be reflected on your next charge, using the same billing method previously setup on this
-						subscription. Please note you can only change tiers once every 30 days. You can always manage your
-						subscriptions on the
-						<router-link to="/subscriptions" class="underline">subscriptions page</router-link>.
-					</p>
-					<div class="flex flex-row-reverse w-full">
-						<button
-							:class="selectedTier !== null ? `` : `opacity-50 cursor-not-allowed`"
-							class="bg-darkBG text-lightButtonText focus:outline-none transform rounded-lg font-bold transition duration-500 ease-in-out hover:bg-opacity-75"
-							style="padding: 0.4rem 1.5rem"
-							:disabled="selectedTier === null"
-							@click="switchTier"
-						>
-							<SpinnerIcon v-if="isLoading" class="mx-2 p-1" />
-							<span v-else class="font-sans" style="font-size: 0.95rem"> Confirm change </span>
-						</button>
-					</div>
+				<article v-if="step === 1" class="flex flex-col items-center">
+					<EditTierConfirm
+						:author="props.author"
+						:sub="props.sub"
+						:selected-tier="selectedTier"
+						:selected-period="selectedPeriod"
+						:payment-profile="paymentProfile"
+						@close="$emit(`close`)"
+					/>
 				</article>
 			</div>
 		</section>
